requirements.txt
colorama==0.4.3
pylint==2.6.0
pyperclip-1.8.2
